// frontend/libs/health-department/domain/src/lib/enums/alert.ts
export enum Alert {
  FIRST_CHARACTERISTIC_SYMPTOM = 'FIRST_CHARACTERISTIC_SYMPTOM',
  DIARY_ENTRY_MISSING = 'DIARY_ENTRY_MISSING',
  INCREASED_TEMPERATURE = 'INCREASED_TEMPERATURE',
  MISSING_DETAILS_INDEX = 'MISSING_DETAILS_INDEX',
  MISSING_DETAILS_CONTACT = 'MISSING_DETAILS_CONTACT',
  QUARANTINE_ENDING = 'QUARANTINE_ENDING',
  ENCOUNTER_IN_QUARANTINE = 'ENCOUNTER_IN_QUARANTINE',
  INITIAL_CALL_OPEN_INDEX = 'INITIAL_CALL_OPEN_INDEX',
  INITIAL_CALL_OPEN_CONTACT = 'INITIAL_CALL_OPEN_CONTACT',
  SUSPICUOUS_SYMPTOM = 'SUSPICUOUS_SYMPTOM',
  UNDEFINED = 'UNDEFINED',
}